pyproject.toml
[tool.poetry]
name = "gtCPACdump"
version = "0.1.0"
description = "Ghost Trick cpac_2d.bin extractor"
license = "MPL-2.0"
authors = ["Slavfox <slavfoxman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Pillow = "^7.0.0"
numpy = "^1.18.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
